+++ example/foglet.html
@@ -62,11 +62,11 @@
             <div class='row' >
 
               <div class='col-md-6' >
-                <h2> Overlay </h2>
+                <h2> RPS </h2>
                 <div id='mainGraph'> </div>
               </div>
               <div class='col-md-6' >
-                <h2> RPS </h2>
+                <h2> OVERLAY </h2>
                 <div id='bisGraph'> </div>
               </div>
             </div>
