+++ src/network/network.js
@@ -4,8 +4,6 @@
 const Communication = require('./communication/communication.js');
 // Signaling
 const Signaling = require('./signaling/signaling.js');
-// debug
-const debug = require('debug')('foglet-core:network');
 
 /**
  * A network is composed of a RPS/Overlay, a signaling part and a communication part
@@ -13,9 +11,8 @@ const debug = require('debug')('foglet-core:network');
 class Network {
   constructor (network, options) {
     this.network = network;
-    this.signaling = new Signaling(network.rps, options.signaling);
-    this.communication = new Communication(network.rps, options.protocol);
-    debug('Network initialized for protocol: '+options.protocol);
+    this.signaling = new Signaling(network, options.signaling);
+    this.communication = new Communication(network, options.protocol);
   }
 }
